+++ src/docs/components/arrows/index.js
@@ -1,0 +1,2 @@
+export { default as InvertedArrow } from './invertedArrowDoc';
+export { default as SimpleArrow } from './simpleArrowDoc';
